% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckhand.R
\name{deckhand}
\alias{deckhand}
\title{Grid layout pages}
\usage{
deckhand(css = NULL, page_size = NULL, ...)
}
\arguments{
\item{css}{Additional CSS files to include}

\item{page_size}{Set page size}

\item{...}{Additional arguments passed to pagedown}
}
\description{
Grid layout pages
}
