% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{deckhand_config_to_css}
\alias{deckhand_config_to_css}
\title{Convert deckhand configuration YAML to a CSS file}
\usage{
deckhand_config_to_css(config_yml = NULL, write = TRUE)
}
\arguments{
\item{config_yml}{A YAML file with deckhand configuration settings}

\item{write}{Whether to write the CSS to a file}
}
\value{
If \code{write = TRUE} invisibly returns the path of the CSS file,
otherwise it will print the CSS to the console and return it as a
character vector.
}
\description{
The deckhand R Markdown formats use CSS to manage page layout, to assist
those with limited knowledge of CSS there is a simple YAML format for
managing general settings.
}
