% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new_config_yml}
\alias{new_config_yml}
\title{Create deckhand configuration YAML}
\usage{
new_config_yml(path = NULL, template = NULL)
}
\arguments{
\item{path}{The path of the new file}

\item{template}{The path of a template file}
}
\description{
If \code{path} and \code{template} are \code{NULL} (the default), deckhand will copy the
default YAML file as \verb{_deckhand.yml} in directory defined by \code{here::here()}.
If using RStudio the file will be opened as a new source document for you
to edit.
}
